+++ packages/sdk-ui/src/index.ts
@@ -1,7 +1,8 @@
 import { First } from './first/first';
+import { ItemView } from './item-view/item-view';
 
 const testMe = () => {
   console.log('Hi, I have been tested.');
 };
 
-export { First, testMe };
+export { First, testMe, ItemView };