+++ packages/sdk-install-modal-web/CHANGELOG.md
@@ -7,8 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.12.1]
-### Uncategorized
-- align version with sdk
+### Added
+- Force align package version to sdk
 
 ## [0.12.0]
 ### Added